requirements.txt
lxml==4.9.3
Pillow==10.0.1
python-docx==1.0.0
typing_extensions==4.8.0
